requirements.txt
steem
websocket-client